@@ -1,11 +1,12 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import './scss/index.scss'
+import React from "react"
+import ReactDOM from "react-dom/client"
+import "./scss/index.scss"
 import App from "./components/App/App"
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById("root"))
+
+root.render(
 	<React.StrictMode>
-		<App/>
-	</React.StrictMode>,
-	document.getElementById('root')
+		<App />
+	</React.StrictMode>
 )
